requirements.txt
Django
django-bootstrap-v5
djangorestframework
django-cors-headers
requests